Make ServerHandler a non nested class of HttpServerImpl
