+++ src/main/java/io/vertx/core/http/impl/ServerConnection.java
@@ -220,7 +220,7 @@ public class ServerConnection extends Http1xConnectionBase implements HttpConnec
     if (ws != null) {
       return ws;
     }
-    HttpServerImpl.ServerHandler serverHandler = (HttpServerImpl.ServerHandler) chctx.pipeline().get("handler");
+    ServerHandler serverHandler = (ServerHandler) chctx.pipeline().get("handler");
     handshaker = serverHandler.createHandshaker(this, chctx.channel(), nettyReq);
     if (handshaker == null) {
       throw new IllegalStateException("Can't upgrade this request");
